Jarred - Made changes to classes
