+++ .idea/libraries/Gradle__com_android_support_constraint_constraint_layout_1_1_3.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.constraint:constraint-layout-1.1.3">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.1.3.aar/0061c6f0a857f82bcb6dcea23908259d/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.1.3.aar/0061c6f0a857f82bcb6dcea23908259d/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.1.3.aar/9534ac507516877bc1614ca509f9bbd4/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.1.3.aar/9534ac507516877bc1614ca509f9bbd4/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
