+++ .idea/libraries/Gradle__com_android_support_test_espresso_espresso_core_3_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test.espresso:espresso-core-3.0.2">
     <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/9a42d0757de9524fb96b76a837b1e5ee/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/9a42d0757de9524fb96b76a837b1e5ee/res" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/2d6ccf44312ea521e3506408ee669a45/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/2d6ccf44312ea521e3506408ee669a45/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
