+++ .idea/libraries/Gradle__com_android_support_test_espresso_espresso_idling_resource_3_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test.espresso:espresso-idling-resource-3.0.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/e0066e551f12fd5388581a9fee777a2a/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/e0066e551f12fd5388581a9fee777a2a/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/27ea16e132d17819cf5e5d38a7a91093/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/27ea16e132d17819cf5e5d38a7a91093/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
