+++ .idea/libraries/Gradle__com_android_support_test_monitor_1_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test:monitor-1.0.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/c8393dbe87ec3e76e9aa0e0ee2bb904c/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/c8393dbe87ec3e76e9aa0e0ee2bb904c/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/62b0bd8f00039065a52f892bafefc2fb/res" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/62b0bd8f00039065a52f892bafefc2fb/jars/classes.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
