+++ .idea/libraries/Gradle__com_android_support_test_runner_1_0_2.xml
@@ -1,8 +1,8 @@
 <component name="libraryTable">
   <library name="Gradle: com.android.support.test:runner-1.0.2">
     <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/0f308eb82746dd8bdb560e5029c67d9f/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/0f308eb82746dd8bdb560e5029c67d9f/jars/classes.jar!/" />
+      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/11e65ae7f67df8aeeb066d38290ca7f1/jars/classes.jar!/" />
+      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/11e65ae7f67df8aeeb066d38290ca7f1/res" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
