--- .idea/libraries/com_android_support_constraint_constraint_layout_1_1_3.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="com.android.support.constraint:constraint-layout-1.1.3">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.1.3.aar/656d190dd25c93a825786c11d838325d/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/constraint-layout-1.1.3.aar/656d190dd25c93a825786c11d838325d/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
