--- .idea/libraries/com_android_support_constraint_constraint_layout_solver_1_1_3_jar.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="com.android.support.constraint:constraint-layout-solver:1.1.3@jar">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support.constraint/constraint-layout-solver/1.1.3/bde0667d7414c16ed62d3cfe993cff7f9d732373/constraint-layout-solver-1.1.3.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
