--- .idea/libraries/com_android_support_test_espresso_espresso_core_3_0_2.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="com.android.support.test.espresso:espresso-core-3.0.2">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/d1639a6f66fe3df5e8c75303abb8d635/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-core-3.0.2.aar/d1639a6f66fe3df5e8c75303abb8d635/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support.test.espresso/espresso-core/3.0.2/433f4f601a3f3835691d80838dd7ff9414931b05/espresso-core-3.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
