--- .idea/libraries/com_android_support_test_espresso_espresso_idling_resource_3_0_2.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="com.android.support.test.espresso:espresso-idling-resource-3.0.2">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/1d668d26efb2da1670a45e4884b18319/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/espresso-idling-resource-3.0.2.aar/1d668d26efb2da1670a45e4884b18319/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support.test.espresso/espresso-idling-resource/3.0.2/6071e2b310c16836ba475a94df377e2f75928546/espresso-idling-resource-3.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
