--- .idea/libraries/com_android_support_test_monitor_1_0_2.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="com.android.support.test:monitor-1.0.2">
-    <CLASSES>
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/964f8c14aa9080edaa2239c6479f6ae5/res" />
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/monitor-1.0.2.aar/964f8c14aa9080edaa2239c6479f6ae5/jars/classes.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support.test/monitor/1.0.2/5764301056100b445e9ba31df7d1b554a55450f6/monitor-1.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
