--- .idea/libraries/com_android_support_test_runner_1_0_2.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="com.android.support.test:runner-1.0.2">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/13c7060d6d7ef8f506e7054c76bbcbfb/jars/classes.jar!/" />
-      <root url="file://$USER_HOME$/.gradle/caches/transforms-1/files-1.1/runner-1.0.2.aar/13c7060d6d7ef8f506e7054c76bbcbfb/res" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.android.support.test/runner/1.0.2/b2ce2342744a66e935dbc4a298259709a4c1f042/runner-1.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
