--- .idea/libraries/com_google_code_findbugs_jsr305_2_0_1_jar.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="com.google.code.findbugs:jsr305:2.0.1@jar">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/com.google.code.findbugs/jsr305/2.0.1/516c03b21d50a644d538de0f0369c620989cd8f0/jsr305-2.0.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
