--- .idea/libraries/net_sf_kxml_kxml2_2_3_0_jar.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="net.sf.kxml:kxml2:2.3.0@jar">
-    <CLASSES>
-      <root url="jar://$APPLICATION_HOME_DIR$/gradle/m2repository/net/sf/kxml/kxml2/2.3.0/kxml2-2.3.0.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
