+++ app/src/main/java/medigram/medigram/Photo.java
@@ -1,4 +1,10 @@
 package medigram.medigram;
 
 public class Photo {
+    private String photoID;
+    private Integer size;
+
+    public void resize(Integer size){
+
+    }
 }
