+++ app/src/main/java/medigram/medigram/Problem.java
@@ -7,10 +7,12 @@ public class Problem {
     private String problemTitle;
     private String description;
     private Date dateStarted;
+    private String bodyLocation;
     private RecordList records;
     private Photo photos;
 
-    public Problem(String problemTitle, String description, Date dateStarted, String bodyLocation, Photo photos){
+    public Problem(String problemTitle, String description, Date dateStarted,
+                   String bodyLocation){
         this.problemTitle = problemTitle;
         this.description = description;
         this.dateStarted = dateStarted;
