+++ app/src/main/java/medigram/medigram/RecordList.java
@@ -1,4 +1,26 @@
 package medigram.medigram;
 
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
 public class RecordList {
+    private List<Record> records = new ArrayList<Record>();
+
+    public void addRecord(String recordTitle, String description,
+                          Date dateStarted, String bodyLocation, Photo photos) {
+
+
+    }
+
+    public void removeRecord(String recordTitle){
+
+    }
+    public Boolean recordExist(String recordTitle){
+
+    }
+
+
+
+
 }
