+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Wed Oct 31 22:20:41 MDT 2018
-sdk.dir=C\:\\Users\\Jarred\\AppData\\Local\\Android\\Sdk
+#Thu Nov 01 16:47:10 MDT 2018
+sdk.dir=/Users/jarred/Library/Android/sdk
